+++ src/cli.ts
@@ -61,7 +61,6 @@ export async function pruneGitHubNotificationsCLI(args: string[]) {
 		args,
 		options: {
 			auth: {
-				default: process.env.GH_TOKEN,
 				type: "string",
 			},
 			bandwidth: {
@@ -102,6 +101,7 @@ export async function pruneGitHubNotificationsCLI(args: string[]) {
 		return;
 	}
 
+	const { auth } = values;
 	const { bandwidth, createdBy, label, lastCommentBy, reason, title, watch } =
 		schema.parse(values);
 	const filters = resolveFilters({
@@ -113,7 +113,7 @@ export async function pruneGitHubNotificationsCLI(args: string[]) {
 	});
 
 	const action = async () =>
-		await pruneGitHubNotifications({ bandwidth, filters });
+		await pruneGitHubNotifications({ auth, bandwidth, filters });
 
 	if (watch) {
 		await runInWatch(action, watch, filters);